update symmetry code structure
